… .travis.yml
… src/full/Agda/Syntax/Translation/ConcreteToAbstract.hs
… src/full/Agda/Syntax/Translation/ReflectedToAbstract.hs
… src/full/Agda/Termination/Monad.hs
… src/full/Agda/TypeChecking/CheckInternal.hs
… src/full/Agda/TypeChecking/CheckInternal.hs-boot
… src/full/Agda/TypeChecking/Constraints.hs
… src/full/Agda/TypeChecking/Constraints.hs-boot
… src/full/Agda/TypeChecking/Conversion.hs
… src/full/Agda/TypeChecking/Conversion.hs-boot
… src/full/Agda/TypeChecking/Coverage.hs
… src/full/Agda/TypeChecking/Datatypes.hs
… src/full/Agda/TypeChecking/DisplayForm.hs
… src/full/Agda/TypeChecking/Eliminators.hs-boot
… src/full/Agda/TypeChecking/Errors.hs
… src/full/Agda/TypeChecking/Errors.hs-boot
… src/full/Agda/TypeChecking/EtaExpand.hs
… src/full/Agda/TypeChecking/Implicit.hs
… src/full/Agda/TypeChecking/Injectivity.hs
… src/full/Agda/TypeChecking/InstanceArguments.hs
… src/full/Agda/TypeChecking/Irrelevance.hs
… src/full/Agda/TypeChecking/Level.hs
… src/full/Agda/TypeChecking/MetaVars.hs
… src/full/Agda/TypeChecking/MetaVars.hs-boot
… src/full/Agda/TypeChecking/MetaVars/Occurs.hs
… src/full/Agda/TypeChecking/Monad/Base.hs
… src/full/Agda/TypeChecking/Monad/Builtin.hs
… src/full/Agda/TypeChecking/Monad/Builtin.hs-boot
… src/full/Agda/TypeChecking/Monad/Caching.hs
… src/full/Agda/TypeChecking/Monad/Closure.hs