The goal of this PR is to make many functions in the Agda codebase effect-polymorphic by generalizing their types from `TCM` to an arbitrary monad `m` satisfying some constraints. In particular, it includes generalized types for the entire pretty-printing pipeline (internal -> abstract and abstract -> concrete), as well as for the conversion checker (i.e. `equalTerms`, `equalTypes`, ...) and the internal double-checker (`checkInternal`). As in the real crusades, I had to pillage many other places throughout the Agda codebase before I got to my final destination.

My main motivation for this refactoring is because I need to call the conversion checker from places where we currently can't, in particular the `ReduceM` monad (see issues #1433, #2732, #3510, #3525). But there are also other reasons why it is a good idea to write monadic code in this style: 
- It statically rules out whole classes of bugs: the types are explicit in which functions can read and/or modify (specific parts of) the typechecking state, throw exceptions and/or warnings, ...
- Having a 'pure' implementation of the `CheckInternal` module allows us to have a separate typechecker for 'core Agda' that is statically guaranteed to not instantiate any metas or create any constraints.
- The constraints on the monad document which effects are actually used by each function. I actually found several lies in the comments (e.g. "Here we use TCM only for the purpose of generating fresh names" when actually it was used for many more things).
- Writing functions against an interface instead of in a specific monad such as `TCM` reduces the coupling of the codebase because not everything has to rely on `TypeChecking.Monad.Base`.

List of new type classes
---------------------------------------
- `MonadConstraint`: creating, solving and modifying constraints
- `MonadMetaSolver`: creating, instantiating, and modifying metavariables
- `MonadAddContext`: (locally) adding things to the context with fresh names
- `MonadInteractionPoints`: creating new interaction points (for the emacs mode)
- `MonadFresh`: generating fresh things (parametrized by the type of the thing)
- `MonadStatistics`: write access to the statistics (for profiling)

TODO:
--------------
- [x] Fix a mysterious internal error that pops up during debug printing
- [x] Benchmarking to make sure this doesn't significantly slow down typechecking
- [x] More things probably