@jespercockx 2019-02-27
Here are the results from running `make library-test` on Agda both before and after this PR. There is a moderate slowdown in overall typechecking (about 9%). The difference is spread kind of evenly between typing, serialization, and parsing, so I'm not sure what to make of it. Any Haskell experts can help with finding out how to avoid this slowdown?

Vanilla Agda @ 12941e52a313af0828d84a61705127fedff5dddf
----------------------------------------------------------------------------------

```
Total                        366,311ms            
Miscellaneous                  3,466ms            
Typing                        46,635ms (171,618ms)
Typing.CheckLHS               24,390ms  (42,088ms)
Typing.CheckLHS.UnifyIndices  17,698ms            
Typing.CheckRHS               37,720ms            
Typing.TypeSig                25,515ms            
Typing.OccursCheck            13,882ms            
Typing.With                    5,776ms            
Serialization                 84,296ms (116,172ms)
Serialization.BinaryEncode    25,990ms            
Serialization.Compress         2,349ms            
Serialization.BuildInterface   2,243ms            
Serialization.Sort             1,294ms            
Parsing                        1,359ms  (17,831ms)
Parsing.OperatorsExpr         13,761ms            
Parsing.OperatorsPattern       2,710ms            
Scoping                       13,103ms  (16,210ms)
Scoping.InverseScopeLookup     3,106ms            
Termination                    7,654ms  (11,518ms)
Termination.Compare            1,937ms            
Termination.RecCheck           1,414ms            
Termination.Graph                511ms            
Highlighting                  10,994ms            
Positivity                     6,833ms            
Coverage                       4,960ms            
Injectivity                    2,916ms            
DeadCode                       2,760ms            
Import                           548ms            
Deserialization                  324ms            
ProjectionLikeness               143ms            
ModuleName                        10ms            

 371,939,415,384 bytes allocated in the heap
  44,886,106,256 bytes copied during GC
   1,167,892,464 bytes maximum residency (68 sample(s))
      18,609,776 bytes maximum slop
            1564 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     190263 colls,     0 par   56.842s  56.859s     0.0003s    0.1564s
  Gen  1        68 colls,     0 par   95.823s  96.050s     1.4125s    3.9330s

  INIT    time    0.000s  (  0.001s elapsed)
  MUT     time  213.649s  (213.830s elapsed)
  GC      time  152.665s  (152.910s elapsed)
  EXIT    time    0.453s  (  0.453s elapsed)
  Total   time  366.768s  (367.194s elapsed)

  %GC     time      41.6%  (41.6% elapsed)

  Alloc rate    1,740,893,595 bytes per MUT second

  Productivity  58.4% of total user, 58.4% of total elapsed


real	6m7,253s
user	6m2,661s
sys	0m4,159s
```

Modified Agda @ 86120bd8037b248f2036a75da425508cdc53bfab
---------------------------------------------------------------------------------------
```
Total                        397,934ms            
Miscellaneous                  3,189ms            
Typing                        46,711ms (185,510ms)
Typing.CheckRHS               56,975ms            
Typing.TypeSig                32,683ms            
Typing.CheckLHS               15,630ms  (29,258ms)
Typing.CheckLHS.UnifyIndices  13,627ms            
Typing.With                   10,002ms            
Typing.OccursCheck             9,878ms            
Serialization                 83,236ms (127,857ms)
Serialization.BinaryEncode    25,525ms            
Serialization.Sort            11,604ms            
Serialization.BuildInterface   5,589ms            
Serialization.Compress         1,901ms            
Parsing                        1,468ms  (20,474ms)
Parsing.OperatorsExpr         15,830ms            
Parsing.OperatorsPattern       3,175ms            
Scoping                       12,927ms  (15,985ms)
Scoping.InverseScopeLookup     3,057ms            
Highlighting                  12,133ms            
Termination                    7,316ms  (11,320ms)
Termination.Compare            2,190ms            
Termination.RecCheck           1,369ms            
Termination.Graph                444ms            
Positivity                     9,422ms            
DeadCode                       4,563ms            
Coverage                       4,232ms            
Injectivity                    1,891ms            
Import                           861ms            
Deserialization                  317ms            
ProjectionLikeness               159ms            
ModuleName                        13ms            

 364,690,160,784 bytes allocated in the heap
  45,661,676,640 bytes copied during GC
   1,139,596,288 bytes maximum residency (67 sample(s))
      13,153,800 bytes maximum slop
            1588 MB total memory in use (23 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     209504 colls,     0 par   61.744s  61.746s     0.0003s    0.1701s
  Gen  1        67 colls,     0 par   105.829s  105.834s     1.5796s    3.6685s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time  230.362s  (230.449s elapsed)
  GC      time  167.573s  (167.580s elapsed)
  EXIT    time    0.393s  (  0.393s elapsed)
  Total   time  398.329s  (398.422s elapsed)

  %GC     time      42.1%  (42.1% elapsed)

  Alloc rate    1,583,117,764 bytes per MUT second

  Productivity  57.9% of total user, 57.9% of total elapsed


real	6m38,475s
user	6m35,613s
sys	0m2,768s
```