@UlfNorell 2019-02-27
My guess is that the problem is carrying along monad dictionaries everywhere, where before you where running in `TCM` which is just a thin layer on top of `IO`.

Try adding a `SPECIALIZE` pragma to every single function you overloaded (although I don't think I've ever managed to get any performance out of one of those).