@nad 2019-05-09
Compilation times on my seven year old laptop, using GHC 8.6.5 with `-j3`, for adcd57d5a952b023a3c055d5d0310b479def1738:
* Without the optimisations:
  ```
  real    15m1,526s
  user    23m32,134s
  sys     5m38,682s
  ```
* With the optimisations:
  ```
  real    22m55,709s
  user    35m4,065s
  sys     8m47,846s
  ```

For `make library-test` (with the default RTS settings, i.e. `-H1G -M1.5G`) I get the following results:
* Without optimisations:
  ```
  Total                        267,646ms
  Miscellaneous                  4,656ms
  Typing                        28,461ms (136,235ms)
  Typing.CheckRHS               43,088ms
  Typing.TypeSig                25,408ms
  Typing.CheckLHS               13,575ms  (24,605ms)
  Typing.CheckLHS.UnifyIndices  11,030ms
  Typing.OccursCheck             8,254ms
  Typing.With                    6,416ms
  Serialization                 47,493ms  (70,416ms)
  Serialization.BinaryEncode    16,895ms
  Serialization.Compress         2,687ms
  Serialization.BuildInterface   2,643ms
  Serialization.Sort               696ms
  Parsing                        1,383ms  (15,222ms)
  Parsing.OperatorsExpr         11,150ms
  Parsing.OperatorsPattern       2,688ms
  Highlighting                  11,367ms
  Scoping                        6,960ms   (9,650ms)
  Scoping.InverseScopeLookup     2,689ms
  Positivity                     5,990ms
  Termination                    1,767ms   (4,124ms)
  Termination.Compare            1,293ms
  Termination.RecCheck             790ms
  Termination.Graph                273ms
  Coverage                       3,336ms
  DeadCode                       2,389ms
  Import                         2,353ms
  Injectivity                    1,477ms
  Deserialization                  284ms
  ProjectionLikeness               128ms
  ModuleName                        12ms

     2,327,593,132 bytes allocated in the heap
    13,978,823,916 bytes copied during GC
       553,705,972 bytes maximum residency (29 sample(s))
         2,412,044 bytes maximum slop
               528 MB total memory in use (0 MB lost due to fragmentation)

                                       Tot time (elapsed)  Avg pause  Max pause
    Gen  0      1852 colls,     0 par   31.784s  31.937s     0.0172s    0.2318s
    Gen  1        29 colls,     0 par   18.950s  19.074s     0.6577s    4.6357s

    INIT    time    0.000s  (  0.000s elapsed)
    MUT     time  214.392s  (218.419s elapsed)
    GC      time   50.734s  ( 51.011s elapsed)
    EXIT    time    0.000s  (  0.000s elapsed)
    Total   time  265.126s  (269.430s elapsed)

    %GC     time       0.0%  (0.0% elapsed)

    Alloc rate    10,856,728 bytes per MUT second

    Productivity  80.9% of total user, 81.1% of total elapsed


  real    4m29,486s
  user    4m25,128s
  sys     0m3,021s
  ```
* With optimisations:
  ```
  Total                        223,952ms
  Miscellaneous                  4,788ms
  Typing                        22,858ms (103,447ms)
  Typing.CheckRHS               28,374ms
  Typing.CheckLHS               11,069ms  (20,531ms)
  Typing.CheckLHS.UnifyIndices   9,461ms
  Typing.TypeSig                19,491ms
  Typing.OccursCheck             6,750ms
  Typing.With                    5,441ms
  Serialization                 46,682ms  (66,003ms)
  Serialization.BinaryEncode    12,855ms
  Serialization.Compress         3,142ms
  Serialization.BuildInterface   2,663ms            
  Serialization.Sort               660ms            
  Parsing                        1,249ms  (13,882ms)
  Parsing.OperatorsExpr          9,980ms            
  Parsing.OperatorsPattern       2,652ms            
  Highlighting                  10,886ms            
  Scoping                        5,428ms   (8,488ms)
  Scoping.InverseScopeLookup     3,060ms            
  Positivity                     4,995ms            
  Termination                    1,439ms   (3,091ms)
  Termination.Compare            1,013ms            
  Termination.RecCheck             489ms            
  Termination.Graph                149ms            
  Coverage                       2,916ms            
  DeadCode                       2,536ms            
  Import                         1,352ms            
  Injectivity                    1,168ms            
  Deserialization                  280ms            
  ProjectionLikeness               101ms            
  ModuleName                        11ms            

     3,805,222,776 bytes allocated in the heap
    13,568,998,588 bytes copied during GC
       528,303,532 bytes maximum residency (28 sample(s))
         2,169,428 bytes maximum slop
               503 MB total memory in use (1 MB lost due to fragmentation)

                                       Tot time (elapsed)  Avg pause  Max pause
    Gen  0      1793 colls,     0 par   30.749s  30.906s     0.0172s    0.2605s
    Gen  1        28 colls,     0 par   13.804s  13.922s     0.4972s    1.6573s

    INIT    time    0.000s  (  0.000s elapsed)
    MUT     time  177.283s  (180.569s elapsed)
    GC      time   44.553s  ( 44.828s elapsed)
    EXIT    time    0.000s  (  0.000s elapsed)
    Total   time  221.836s  (225.397s elapsed)

    %GC     time       0.0%  (0.0% elapsed)

    Alloc rate    21,464,157 bytes per MUT second

    Productivity  79.9% of total user, 80.1% of total elapsed


  real    3m45,448s
  user    3m41,837s
  sys     0m2,794s
  ```

My build times are considerably longer than @jespercockx's, but the library test completes more quickly. I think @jespercockx has newer, more performant hardware, but I'm using a newer version of GHC and (unlike @jespercockx) a 32-bit kernel. (And we haven't tested exactly the same version of Agda, and perhaps not exactly the same version of the standard library.) An apples-to-apples comparison where only one variable at a time is tweaked might be interesting.